unnapropiate comment removal
at the start of this file was a commentary that can be misundertood. that were not my words, it was a lyrics from 
"dj sound - hoish ass n*ggaz".
i felt writting whatever i just listened in the moment and i just forgot i had it there lol sorry wasnt my intention to sound racist.